retention mode and sleep params in lib.rs
