+++ src/blocking/worker_pool.rs
@@ -2,9 +2,8 @@ use crate::diesel::r2d2;
 use crate::diesel::PgConnection;
 use crate::error::FangError;
 use crate::executor::Executor;
-use crate::executor::RetentionMode;
-use crate::executor::SleepParams;
 use crate::queue::Queue;
+use crate::{RetentionMode, SleepParams};
 use log::error;
 use log::info;
 use std::collections::HashMap;
